Don't set the Message ID for hello messages
The Message ID field is used by notification messages to identify a
given message.

According with RFC 5036 - section 3.5.1.2.1, malformed LDP PDUs or
messages that are part of the LDP Discovery mechanism are handled by
silently discarding them. With that said, there's no need to set the
Message ID field for hello messages.

This is the behavior adopted by Cisco IOS.