Speed-up the session establishment process
* Send an extra Hello message before attempting to connect to a remote
peer to guarantee that it formed an adjacency with us as well;

* Don't wait for the first timeout to send the first Hello message. If
we are playing the passive role, the remote peer will immediately
establish a TCP connection with us and then send an Initialization
message. The problem is that we may have not formed an adjancency with
the remote peer yet. In this case, we would close the TCP connection
and then send a notification message (No Hello), which would trigger the
backoff exponential timer in the remote peer. In practice this problem
doesn't occurs because most implementations use the first optimization
mentioned above.

Note: the Cisco IOS implementation only sends an extra hello message
_after_ the TCP connection with the passive peer is established and
_before_ sending the Initialization message. Without the previous
commit, we wouldn't receive the extra Hello message from IOS, which
would considerably slow down the session establishment process.