Advertise the implicit-null label for routes attached to loopback interfaces
This is necessary in order to guarantee PHP behavior when we are in the
edge of an MPLS network.

Note: without this patch LDP won't work if the 'fib-update no' command is
set. The LDP neighbors will send labeled packets to our loopback address
which will be discarded since the daemon is not feeding the kernel.