Don't assign labels for BGP routes.
Although RFC 5036 is not explicit about this, LDP should not assign
labels for BGP routes. Doing that would be very resource consuming in
some scenarios and unnecessary. The goal is generally only to establish
LSPs among all PEs in the AS since LDP is not used as an end in itself
but as a means to implement advanced solutions like MPLS L2/L3 VPNs. Some
implementations (e.g. JunOS) go further and only assign labels for /32
loopback routes advertised in the IGP.

If Inter-AS LSPs are necessary, BGP itself should be used for distributing
IPv4 labeled routes (e.g. option C. of section 10 in RFC 4364).