Remove lo protection.
There's no need to protect the 127/8 network on ldpd since this network
is filtered before being sent to lde.

If we receive a label mapping for this network, it won't be installed
because lde has no nexthop for it, and thus the code will always fall
into the LMp.13 case of the RFC "Receive Label Mapping" algorithm:
the mapping will be recorded but not used.