Remove tracking of which interfaces joined which multicast addresses.
All this complexity was inherited from ospfd but it's not necessary
in ldpd.

In ldpd we don't have the concept of different areas or instances,
so there's no risk of one interface trying to join the same multicast
group more than once. In addition to that, ldpd interfaces only join
one single multicast group (224.0.0.2), which makes things trivial.